Move discardChanges git method to apply.ts
We use the first argument of the git command as the name of the file
