+++ app/src/lib/git/apply.ts
@@ -4,10 +4,10 @@ import {
   WorkingDirectoryFileChange,
   AppFileStatusKind,
 } from '../../models/status'
-import { DiffType } from '../../models/diff'
+import { DiffType, ITextDiff, DiffSelection } from '../../models/diff'
 import { Repository, WorkingTree } from '../../models/repository'
 import { getWorkingDirectoryDiff } from './diff'
-import { formatPatch } from '../patch-formatter'
+import { formatPatch, formatPatchToDiscardChanges } from '../patch-formatter'
 
 export async function applyPatchToIndex(
   repository: Repository,
@@ -99,4 +99,40 @@ export async function checkPatch(
   }
 
   return true
+}
+
+/**
+ * Discards the local changes for the specified file based on the passed diff
+ * and a selection of lines from it.
+ *
+ * When passed an empty selection, this method won't do anything. When passed a
+ * full selection, all changes from the file will be discarded.
+ *
+ * @param repository The repository in which to update the working directory
+ *                   with information from the index
+ *
+ * @param filePath   The relative path in the working directory of the file to use
+ *
+ * @param diff       The diff containing the file local changes
+ *
+ * @param selection  The selection of changes from the diff to discard
+ */
+export async function discardChangesFromSelection(
+  repository: Repository,
+  filePath: string,
+  diff: ITextDiff,
+  selection: DiffSelection
+) {
+  const patch = formatPatchToDiscardChanges(filePath, diff, selection)
+
+  if (patch === null) {
+    // When the patch is null we don't need to apply it since it will be a noop.
+    return
+  }
+
+  const args = ['apply', '--unidiff-zero', '--whitespace=nowarn', '-']
+
+  await git(args, repository.path, 'discardChangesFromSelection', {
+    stdin: patch,
+  })
 }
