+++ app/src/lib/git/index.ts
@@ -35,4 +35,3 @@ export * from './rebase'
 export * from './format-patch'
 export * from './worktree'
 export * from './tag'
-export * from './discard-changes'
